requirements.txt
openai-whisper
sounddevice
numpy
PyAudio
tiktoken
torch
tqdm
more-itertools
triton>=2.0.0;platform_machine=="x86_64" and sys_platform=="linux" or sys_platform=="linux2"
